+++ OpenTweet/Models/Timeline.swift
@@ -30,29 +30,3 @@ struct Thread {
   let root: Tweet
   let replies: [Tweet]
 }
-
-struct Tweet: Decodable, Equatable {
-  let id: String
-  let author: String
-  let content: String
-  let date: Date
-  let avatar: URL?
-  let inReplyTo: String?
-  
-  var dateDisplay: String {
-    let formatter = DateFormatter()
-    formatter.dateStyle = .medium
-    formatter.timeStyle = .short
-
-    return formatter.string(from: date)
-  }
-  
-  var threadType: ThreadType {
-    return inReplyTo == nil ? .root : .reply
-  }
-}
-
-enum ThreadType {
-  case root
-  case reply
-}
